Add a computed value of TTL (Time To Land) for each commit indexed
This patch introduces the concept of TTL (Time To Land) that
is a time delta between the author and committer date.

In a sane context where commits are "amended" after each
necessarily modification requests before being merged
this TTL value can be interpreted as the delay of approval
of a patch or its "time to land".

This commit also adds a method to compute statistics around
this TTL value.

A rebuild of the ElasticSearch DB is needed after this patch
has landed.

Change-Id: Iddde498fd824d5d8262ea29f53c71d7b7eb79513