Use multiprocessing module for git commit extraction
Iterating over all commits of projects can be long
for big projects. This patch leverage the use of the
multicore for commit extracting.

By default use host amount of CPU - 1. But can
be fully controlled by the extract-worker arguments.

A follow will be to profile the commit extraction
function to identify the potential bootlenecks.

Also this patch fixes:
- Potential error at str decoding to utf-8
- UI - author amounts not well computed

Change-Id: I90ba31a05885651120f118e7b5c86424207518bd