Skip Packaging-deb by default
Even if the Packaging-deb repositories can be indexed
accuratly its layout introduce some not accurate stats.
Indeed debian/ is applied on top of the upstream project
branches and if we take the example of deb-nova repo its
almost impossible to extract the real contributors to the
deb packaging of Nova because they are mixed with Nova
contributors.

Change-Id: I6162bc3af5ca184fc6bacf5f4739d1d876204aee