Expose querying metadata through the commits.json endpoint
The metadata argument can be used to limit results to
commit that match one or more metadata. For instance
using &metadata=close-bug=10,need-qa=true will limit
to commit that have in the commit message:

close-bug: 10
need-qa: true

Using &metada=close-bug=* will return commits that contains
the metadata close-bug no matter the value.

RepoXplorer can be queried like that for instance:
http://repoxplorer-url/commits.json?pid=test&metadata=close-bug=10

This usage will be detailed in the documentation in
a follow up patch.

Change-Id: Icead086826e422c0da7e0f05eed37c84985814ab