Add a user backend in EL
Contributors index now look for user definitions
in ElasticSearch. The user backend can be populated
by an external process using its API 'users/'.

Contributors definition will always take precedence
over the user backend.

For now its real purpose is if an authentication is
placed in front of repoXplorer and user data (emails)
can be fetched (like for OAuth2 or OpenID) then it is
handy to fill the user backend.

Change-Id: Ibea7e4276fa886a7125fa4ef225a4206d04c0804