Reserve dynamics field name of tag doc type too
This is a workaround as the behavior of EL and mapping
is applied to a doc type but it seems here this
does not apply correctly when a dynamic date field is detected
as string type for the commit doc type it failed to be indexed
due to the tag doc type mapping that contain a date field of date
type. More investigation will be needed.

This patch simply does not index dynamics detected date field.

Change-Id: Ib7d5cb61211d2b4431e8e9412ccf047c474fbce2