Update patch set 5
Patch Set 5:

Nice thank you for that patch that closed to be good to merged.
The code was working almost as expected (even the toggle button).
I submitted a new patchset (5) with a proposal to stack repos and branches in the boxes. Fell free to change/revert or accept it.
Also look at the comments on the PS4.
Furthermore I used to put the semicolon even if it is not mandatory with the JS but it seems to be a good practice (and my emacs linter is happy :) ) so could you add them too.

Patch-set: 5