Fill the info box for the project page using the API
Previously, we were using Maki templating to fill the info box in
project.html.

We are now adding a get_infos JS call, that will be later extended
to include data for other pages (contributor and group), and making
the info box use the data generated from that call.

TG-594
TG-595

Change-Id: I5621c2c80f6439c8a3a757867ef2d716e191c0fd