Update patch set 7
Patch Set 7: Code-Review+2

(1 comment)

For me it is good to be merged. I did a change on your last PS because when I did some testing I figured out some inaccurate results. This was due to the sanitize author function that by default cut the list to 25 id ... Let have a look to my last changes and let me know if that is ok for you.

Patch-set: 7
Label: Code-Review=+2