ui: project pages: move top authors build to js
Furthermore add an ErrorController to build custom
json response for 404 error. That way we can return to
the user the detail string passed to the abort method.

The project page is now able to display a message box
telling about the API 404 reason. For instance an
unknown cid or tid.

Change-Id: I9a6b9477a80323961ab9e8dc97ff52fc9f0eb873