Update patch set 4
Patch Set 4:

Thanks for the review. Yes I want to work on a following patch to have a "display all" button/link in the top listing that will refresh the top box with all results. So by default only 10 items are retrieved from the API and the user can display all (api is re-requested with limit=-1) and the box is refreshed.

Patch-set: 4
Label: Code-Review=0