Fix get_metadata_keys slow on large dataset
The metadata endpoint was slow on large dataset. It was
due to the fact we search metadata keys accross all
commits returned by the filter. So on large projects
like the linux kernel the search was perform accross more
than 700000 commits. As we want to expose to the user the most
common keys then looking to only a limited set of commits make
sense.

Now only the last 10000 commits are queried so the endpoint is
by far faster.

Finally the UI no longer return the hit amount for each metadata
because the value is now inaccurate.

Change-Id: I7ff4dfdb4e30db2128e35f489bb3454d8f3af608