Update patch set 1
Patch Set 1: Workflow-1

Tested live but does not seems to fix the problem. The indexer process start to eat most of the available ram on the node in the following situation: linux kernel repo indexed, and linux-audit/audit-kernel is going to be indexed -> audit-kernel is a mirror of kernel so indexer want to update ~ 700000 commits so depending on the host condition we run out of memory.

Patch-set: 1
Reviewer: Gerrit User 6 <6@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Workflow=-1