Fully include the last day of dto
Unit of date bounces are day based, but the conversion to epoch of
a date such as 2017-08-16 (w/o hours/minutes) is 2017-08-16 00:00 AM.
The full day is not included, that is normal but not expected in our
use case.

To include the full day automatically add a complete day 24*3000s
after the dto conversion.

Change-Id: I8ffac81a919fa2880b8c3024ca4d55b8a93a51ee