Fix display of top new authors by commit
Previously, this did not work when we wanted to display new authors
from a specific set of dates, just new authors from now. The updated
code changes the "since" box meaning slightly:

When there is a "from" and "to" date, it displays the new authors
comparing the "from - since -> from " period to the "from -> to"
period. So for example, if from: 2018-01-01, to: 2018-03-01, and
since: 1 month, if will compare the periods:

- 2017-12-01 -> 2018-01-01, and
- 2018-01-01 -> 2018-03-01

to get the new authors.

TG-1750

Change-Id: I73bf7a7fff52976075e6f127e1ed146719288356