Add index-tags attribute in project-template definition
This attribute is a boolean. By default set to true but if false
then the indexer won't index tags but also will make sure that
existing tags are wipped from the DB.

For packaging projects like Fedora or RDO, it does not make sense
to index tags of every repositories but instead have only user
defined releases.

Change-Id: I78ec2f5e0f33e99b6ea7cc4e8c532c588c7704bf