Add caching strategy for projects structures
Two projects stuctures are needed for most of the API calls but
rebuilt at each request adding some delay to be computed.
Some tests with a big project definition (fedora distgit ~ 30000 repos)
shown that the impact was big by adding ~10 seconds to compute.

This patch adds a strategy to cache both structures. Thanks
to that we are ~ 5x faster than before.

Change-Id: Ib1d0c88e696b3147fe27cb6b9953787ce8f795d9