Wip - add first meta_ref fundations
Idea is to add a meta ref to commits (in the DB) for projects
that specify the use of the meta_ref (no implemented yet) but
for Fedora for example. This meta_ref will be used when
we want to query stats for such projects. So the query will
be simple as we will filter commit by the meta_ref only and not
will all refs (for Fedora it is ~ 30000 refs). 30000 or in
the EL query is not adviced at all and is very slow.

Change-Id: I3cd1bfce753ec535068a4f5161d7b4048ccf47f8