Meta ref implementation to ease deal with bigger projects
A flag: 'meta_ref: True' can be added to project description.
If the tag is set then indexer will add an additional ref
in 'repos' attribute in the EL DB for commits part of the
project.

At query, if the query is full project scope (w/o inc_repos, or
filter by paths) then the meta_ref is used. This simply
the OR query by using only one 'repos' clause.

This is to mitigate performance degradation when amount
of refs for a project can be high, like for Fedora
( ~ 30000 refs).

Previously indexed commit won't get automatically that
meta_ref if the project definition is changed. So
as of now, the project must be fully re-indexed.

Change-Id: I3cd1bfce753ec535068a4f5161d7b4048ccf47f8