Make better usage of the infos endpoint
projects_amount and repos_amount metric comes from
the infos endpoint for every kind of resources project,
group, contributor.

We no longer use the contributor or group endpoint to
fetch that data.

Furthermore, and it was the main purpose of that patch,
project info box "Repository refs" is based on filters.

Change-Id: Ifbfc57c303f383334aaa36837a5b66d10d0252c3