Projects definition stored in elasticsearch
The current implementation have limitations in terms of
performances and scaling. Big deployement can experiences
high memory usage issue in the indexer and rest process.

This patch dumps the projects definition in the ElasticSearch
instance. Flat files are read then dumped in the ELS.

This add an option to deactivate the dump of the definition
in EL. The struct build (prior to the dump) can take
some time and one can want to trigger it externaly

TODO:
- perform manual tests

Change-Id: Ief98c8d2a14e49fba3c94a2136ee4891d74a2a15