Projects definition stored in elasticsearch
The current implementation have limitations in terms of
performances and scaling. Big deployement can experience
high memory usage issue in the indexer and API process.

This patch dumps the projects definition in the ElasticSearch
instance. Flat files are read then dumped in the ELS.

The projects definition is dumped only at repoxplorer-indexer
calls. A new option 'refresh-projects-index' has been added, that
can be used by a CD process to refresh the definition in ELS.

Other advantage of having project definitions in ELS is
that will ease the writting of a full CRUD REST endpoint to
manage project definitions.

This patch also contains some additionnal code refactoring.
This work will be extended to the group definition asap.

Change-Id: Ief98c8d2a14e49fba3c94a2136ee4891d74a2a15